handle .bcf negative QUAL properly

+++ plink_data.c
@@ -9144,7 +9144,7 @@ int32_t bcf_to_bed(char* bcfname, char* outname, char* outname_end, int32_t miss
   uint32_t gt_idx = 0;
   uint32_t marker_ct = 0;
   uint32_t umm = 0;
-  uint32_t vcf_min_qualf_bits = 0;
+  int32_t vcf_min_qualf_bits = 0;
   int32_t retval = 0;
   char missing_geno = *g_missing_geno_ptr;
   uint32_t bcf_var_header[8];
@@ -9469,7 +9469,7 @@ int32_t bcf_to_bed(char* bcfname, char* outname, char* outname_end, int32_t miss
       goto bcf_to_bed_marker_skip;
     }
     if (check_qual) {
-      if ((bcf_var_header[5] == 0x7f800001) || (bcf_var_header[5] < vcf_min_qualf_bits)) {
+      if ((bcf_var_header[5] == 0x7f800001) || (((int32_t)bcf_var_header[5]) < vcf_min_qualf_bits)) {
         goto bcf_to_bed_marker_skip;
       }
     }
